test: add package parsing unit tests
