+++ app/api/scans/__tests__/packageParsing.test.ts
@@ -1,0 +1,75 @@
+import { describe, expect, it } from "vitest";
+
+import { parsePackageContent } from "../packageParsing";
+
+type ParsedPackage = ReturnType<typeof parsePackageContent>[number];
+
+const sortPackages = (packages: ParsedPackage[]) =>
+  packages
+    .slice()
+    .sort(
+      (a, b) => a.name.localeCompare(b.name) || a.version.localeCompare(b.version),
+    );
+
+describe("parsePackageContent", () => {
+  it("parses package.json dependencies", () => {
+    const content = JSON.stringify({
+      dependencies: {
+        "left-pad": "^1.3.0",
+        lodash: "4.17.21",
+      },
+      devDependencies: {
+        typescript: "~5.2.0",
+        invalid: "not-a-version",
+      },
+    });
+
+    const result = parsePackageContent(content);
+
+    expect(sortPackages(result)).toEqual(
+      sortPackages([
+        { name: "left-pad", version: "1.3.0", isDirect: true },
+        { name: "lodash", version: "4.17.21", isDirect: true },
+        { name: "typescript", version: "5.2.0", isDirect: true },
+      ]),
+    );
+  });
+
+  it("parses package-lock v3 without name fields", () => {
+    const content = JSON.stringify({
+      lockfileVersion: 3,
+      packages: {
+        "": {
+          name: "root",
+          dependencies: {
+            "left-pad": "^1.3.0",
+            "@scope/pkg": "^2.0.0",
+          },
+        },
+        "node_modules/left-pad": {
+          version: "1.3.0",
+        },
+        "node_modules/@scope/pkg": {
+          version: "2.1.0",
+        },
+        "node_modules/left-pad/node_modules/sub": {
+          version: "0.1.0",
+        },
+      },
+    });
+
+    const result = parsePackageContent(content);
+
+    expect(sortPackages(result)).toEqual(
+      sortPackages([
+        { name: "left-pad", version: "1.3.0", isDirect: true },
+        { name: "@scope/pkg", version: "2.1.0", isDirect: true },
+        { name: "sub", version: "0.1.0", isDirect: false },
+      ]),
+    );
+  });
+
+  it("returns empty for invalid JSON", () => {
+    expect(parsePackageContent("{invalid")).toEqual([]);
+  });
+});
